Search service now requests custom connectors for standard as well as consumption
