+++ libs/services/designer-client-services/src/lib/consumption/search.ts
@@ -15,23 +15,6 @@ export class ConsumptionSearchService extends BaseSearchService {
     );
   }
 
-  public async getAllCustomApiOperations(): Promise<DiscoveryOpArray> {
-    try {
-      const {
-        apiHubServiceDetails: { apiVersion, subscriptionId, location },
-      } = this.options;
-      const uri = `/subscriptions/${subscriptionId}/providers/Microsoft.Web/locations/${location}/apiOperations`;
-      const queryParameters: QueryParameters = {
-        'api-version': apiVersion,
-        $filter: `type eq 'Microsoft.Web/customApis/apiOperations' and ${ISE_RESOURCE_ID} eq null`,
-      };
-      return await this.batchAzureResourceRequests(uri, queryParameters);
-    } catch (error) {
-      console.error(error);
-      return [];
-    }
-  }
-
   public getConsumptionBuiltInOperations(): DiscoveryOpArray {
     const clientBuiltInOperations = getClientBuiltInOperations(true);
     const consumptionBuiltIn: any[] = [
@@ -56,23 +39,6 @@ export class ConsumptionSearchService extends BaseSearchService {
     return Promise.all([this.getAllAzureConnectors(), this.getAllCustomApiConnectors(), this.getConsumptionBuiltInConnectors()]).then(
       (values) => values.flat()
     );
-  }
-
-  public async getAllCustomApiConnectors(): Promise<Connector[]> {
-    try {
-      const {
-        apiHubServiceDetails: { apiVersion, subscriptionId },
-      } = this.options;
-      const uri = `/subscriptions/${subscriptionId}/providers/Microsoft.Web/customApis`;
-      const queryParameters: QueryParameters = {
-        'api-version': apiVersion,
-        $filter: `${ISE_RESOURCE_ID} eq null`,
-      };
-      return await this.getAzureResourceRecursive(uri, queryParameters);
-    } catch (error) {
-      console.error(error);
-      return [];
-    }
   }
 
   public getConsumptionBuiltInConnectors(): Connector[] {
